Enhanced ArrayReverser Program with Input Validation and Clean Output
